+++ civis/ml/_model.py
@@ -24,7 +24,7 @@ except ImportError:
 from civis import APIClient, find, find_one, __version__
 from civis._utils import camel_to_snake
 from civis.base import CivisAPIError, CivisJobFailure
-from civis.compat import FileNotFoundError
+from civis.compat import FileNotFoundError, TemporaryDirectory
 import civis.io as cio
 from civis.futures import ContainerFuture
 
@@ -81,7 +81,7 @@ def _block_and_handle_missing(method):
     return wrapper
 
 
-def _stash_local_dataframe(df, client=None):
+def _stash_local_dataframe(df, template_id, client=None):
     """Store data in a temporary Civis File and return the file ID"""
     # Standard dataframe indexes do not have a "levels" attribute,
     # but multiindexes do. Checking for this attribute means we don't
@@ -90,6 +90,33 @@ def _stash_local_dataframe(df, client=None):
         raise TypeError("CivisML does not support multi-indexed data frames. "
                         "Try calling `.reset_index` on your data to convert "
                         "it into a CivisML-friendly format.")
+    try:
+        if template_id > 9969:
+            return _stash_dataframe_as_feather(df, client)
+        else:
+            return _stash_dataframe_as_csv(df, client)
+    except (ImportError, AttributeError) as exc:
+        if (df.dtypes == 'category').any():
+            # The original exception should tell users if they need
+            # to upgrade pandas (an AttributeError)
+            # # or if they need to install "feather-format" (ImportError).
+            six.raise_from(ValueError(
+                'Categorical columns can only be handled with pandas '
+                'version >= 0.20 and `feather-format` installed.'),
+                exc)
+        return _stash_dataframe_as_csv(df, client)
+
+
+def _stash_dataframe_as_feather(df, client):
+    civis_fname = 'modelpipeline_data.feather'
+    with TemporaryDirectory() as tdir:
+        path = os.path.join(tdir, civis_fname)
+        df.to_feather(path)
+        file_id = cio.file_to_civis(path, name=civis_fname, client=client)
+    return file_id
+
+
+def _stash_dataframe_as_csv(df, client):
     civis_fname = 'modelpipeline_data.csv'
     buf = six.BytesIO()
     if six.PY3:
@@ -837,6 +864,9 @@ class ModelPipeline:
             Note that the index of the :class:`~pandas.DataFrame` will be
             ignored -- use ``df.reset_index()`` if you want your
             index column to be included with the data passed to CivisML.
+            NB: You must install ``feather-format`` if your
+            :class:`~pandas.DataFrame` contains :class:`~pandas.Categorical`
+            columns, to ensure that CivisML preserves data types.
         csv_path : str, optional
             The location of a CSV of data on the local disk.
             It will be uploaded to a Civis file.
@@ -894,7 +924,8 @@ class ModelPipeline:
             raise ValueError('Provide a single source of data.')
 
         if df is not None:
-            file_id = _stash_local_dataframe(df, client=self._client)
+            file_id = _stash_local_dataframe(df, self.train_template_id,
+                                             client=self._client)
         elif csv_path:
             file_id = _stash_local_file(csv_path, client=self._client)
 
@@ -1079,6 +1110,9 @@ class ModelPipeline:
             Note that the index of the :class:`~pandas.DataFrame` will be
             ignored -- use ``df.reset_index()`` if you want your
             index column to be included with the data passed to CivisML.
+            NB: You must install ``feather-format`` if your
+            :class:`~pandas.DataFrame` contains :class:`~pandas.Categorical`
+            columns, to ensure that CivisML preserves data types.
         csv_path : str, optional
             The location of a CSV of data on the local disk.
             It will be uploaded to a Civis file.
@@ -1140,7 +1174,8 @@ class ModelPipeline:
             raise ValueError('Provide a single source of data.')
 
         if df is not None:
-            file_id = _stash_local_dataframe(df, client=self._client)
+            file_id = _stash_local_dataframe(df, self.predict_template_id,
+                                             client=self._client)
         elif csv_path:
             file_id = _stash_local_file(csv_path, client=self._client)