+++ civis/tests/test_parallel.py
@@ -202,7 +202,7 @@ def test_infer(mock_make_factory, mock_job):
         max_submit_retries=0,
         max_job_retries=0,
         hidden=True,
-        remote_backend='civis',
+        remote_backend='sequential',
         **expected)
 
 
@@ -313,7 +313,7 @@ def test_infer_from_custom_job(mock_make_factory):
                        'max_submit_retries': mock.ANY,
                        'max_job_retries': mock.ANY,
                        'hidden': True,
-                       'remote_backend': 'civis'}
+                       'remote_backend': 'sequential'}
     for key in civis.parallel.KEYS_TO_INFER:
         expected_kwargs[key] = mock_script[key]
     mock_make_factory.assert_called_once_with(**expected_kwargs)