+++ CHANGELOG.md
@@ -17,6 +17,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   Require this when using ``pd.Categorical`` types, since CSVs require us to
   re-infer column types, and this can fail. Using feather should also give a
   speed improvement; it reads and writes faster than CSVs and produces smaller files (#200).
+- ``ModelFuture`` objects will emit any warnings which occurred during their
+  corresponding CivisML job (#204)
 
 ### Fixed
 - Restored the pre-v1.7.0 default behavior of the ``joblib`` backend by setting the ``remote_backend``